requirements.txt
beautifulsoup4==4.13.3
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.47
langchain_openai==0.3.9
langchain_pinecone==0.2.3
openai==1.68.2
python-dotenv==1.0.1
Requests==2.32.3
pypdf==5.4.0
streamlit==1.43.2
